requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
PyPDF2
python-docx
gradio
